requirements.txt
Django==6.0.1
psycopg[binary]==3.3.2
django-pgtrigger==4.17.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.21
argon2-cffi==25.1.0
django-ninja==1.5.3
django-ninja-jwt==5.4.3
django-two-factor-auth==1.18.1
pillow==12.1.0
pyotp==2.9.0
hiredis==3.3.0
redis==7.1.0
django-redis==6.0.0
pydantic-extra-types==2.11.0
pydantic[email]==2.12.5
pycountry==24.6.1